models: move Status types into models crate
This sets up a future commit to expose status via an http endpoint. We'd like
to have all types that are used in the public API organized under the `models`
crate to make it easy to share code between the client and server. So this
moves the status types and refactors controllers to use functions that take an
explicit `status` parameter instead of `self`.  There are no behavior changes
here.

This also impacted publications, discovers, and evolutions, since we expose a
few types from those as part of the status. I moved over only the minimal set
of types that are embedded within a status, though it may be necessary to move
more later as we introduce http endpoints for discover and publish operations.
