+++ crates/agent/src/evolution.rs
@@ -1,6 +1,6 @@
 use agent_sql::Capability;
 use itertools::Itertools;
-use schemars::JsonSchema;
+use models::evolutions::EvolvedCollection;
 use serde::{Deserialize, Serialize};
 use sqlx::PgPool;
 use std::collections::BTreeSet;
@@ -104,18 +104,6 @@ impl EvolveRequest {
         }
         Ok(())
     }
-}
-
-#[derive(Serialize, Deserialize, PartialEq, Debug, Clone, JsonSchema)]
-pub struct EvolvedCollection {
-    /// Original name of the collection
-    pub old_name: String,
-    /// The new name of the collection, which may be the same as the original name if only materialization bindings were updated
-    pub new_name: String,
-    /// The names of any materializations that were updated as a result of evolving this collection
-    pub updated_materializations: Vec<String>,
-    /// The names of any captures that were updated as a result of evolving this collection
-    pub updated_captures: Vec<String>,
 }
 
 #[tracing::instrument(skip_all, fields(user_id = %evolution.user_id))]
